完成登录功能/重构代码

+++ src/main/java/cn/alumik/shop/dao/RoleRepository.java
@@ -3,6 +3,5 @@ package cn.alumik.shop.dao;
 import cn.alumik.shop.entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
-
+public interface RoleRepository extends JpaRepository<Role, Integer> {
 }
